+++ waltz-ng/client/applications/app-view.js
@@ -81,12 +81,12 @@ function controller($q,
                     capabilityStore,
                     changeLogStore,
                     databaseStore,
-                    dataFlowStore,
                     dataFlowDecoratorStore,
                     dataTypeUsageStore,
                     entityStatisticStore,
                     historyStore,
                     involvementStore,
+                    logicalFlowStore,
                     orgUnitStore,
                     perspectiveStore,
                     physicalSpecificationStore,
@@ -182,7 +182,7 @@ function controller($q,
 
     function loadThirdWave() {
         const promises = [
-            loadDataFlows(dataFlowStore, id, vm),
+            loadDataFlows(logicalFlowStore, id, vm),
             loadInvolvements($q, involvementStore, id, vm),
             loadAuthSources(authSourcesStore, orgUnitStore, id, vm.organisationalUnit.id, vm),
             loadServers(serverInfoStore, id, vm),
@@ -264,12 +264,12 @@ controller.$inject = [
     'CapabilityStore',
     'ChangeLogStore',
     'DatabaseStore',
-    'DataFlowDataStore',
     'DataFlowDecoratorStore',
     'DataTypeUsageStore',
     'EntityStatisticStore',
     'HistoryStore',
     'InvolvementStore',
+    'LogicalFlowStore',
     'OrgUnitStore',
     'PerspectiveStore',
     'PhysicalSpecificationStore',
